refactored signalling
