+++ go.mod
@@ -4,11 +4,11 @@ go 1.23.2
 
 require (
 	github.com/gorilla/websocket v1.5.3
-	github.com/sirupsen/logrus v1.9.3
+	github.com/labstack/echo/v4 v4.12.0
 )
 
 require (
-	github.com/labstack/echo/v4 v4.12.0 // indirect
+	github.com/golang-jwt/jwt v3.2.2+incompatible // indirect
 	github.com/labstack/gommon v0.4.2 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.20 // indirect
@@ -18,4 +18,5 @@ require (
 	golang.org/x/net v0.24.0 // indirect
 	golang.org/x/sys v0.19.0 // indirect
 	golang.org/x/text v0.14.0 // indirect
+	golang.org/x/time v0.5.0 // indirect
 )
