+++ views/landing.html
@@ -6,7 +6,7 @@
     <title>Chat Room</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link href="http://netdna.bootstrapcdn.com/bootstrap/3.3.1/css/bootstrap.min.css" rel="stylesheet">
-    <link rel="stylesheet" type="text/css" href="css/chatbox.css">
+    <link rel="stylesheet" type="text/css" href="/assets/styles.css">
     <link rel="shortcut icon" type="image/png" href="ui/imgs/favicon.png"/>
 </head>
 <body>
@@ -105,7 +105,7 @@
 </div>
 <script src="http://code.jquery.com/jquery-1.10.2.min.js"></script>
 <script src="http://netdna.bootstrapcdn.com/bootstrap/3.3.1/js/bootstrap.min.js"></script>
-<script src = "client.js"></script>
+<script src = "/assets/client.js"></script>
 </body>
 </html>
 {{ end }}
